+++ src/GravatarHelper.AspNetCore/GravatarHelper.AspNetCore.csproj
@@ -7,10 +7,10 @@
     <PackageLicenseUrl>http://creativecommons.org/publicdomain/zero/1.0/</PackageLicenseUrl>
     <PackageProjectUrl>https://github.com/jkommer/GravatarHelper/</PackageProjectUrl>
     <RepositoryUrl>https://github.com/jkommer/GravatarHelper/</RepositoryUrl>
-    <RepositoryType>git</RepositoryType>    
+    <RepositoryType>git</RepositoryType>
     <PackageRequireLicenseAcceptance>false</PackageRequireLicenseAcceptance>
     <Copyright>Copyright (c) 2012, Johannes Kommer.</Copyright>
-    <PackageTags>aspnetcore aspnetcoremvc Gravatar</PackageTags>  
+    <PackageTags>aspnetcore aspnetcoremvc Gravatar</PackageTags>
   </PropertyGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.AspNetCore.Mvc" Version="1.0.0" />
